requirements.txt
PyQt6==6.11.0
edge-tts==7.2.8
pygame-ce==2.5.7
